+++ Source/EventFlow/Commands/ICommand.cs
@@ -20,6 +20,8 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System.Threading;
+using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Core;
 
@@ -27,13 +29,22 @@ namespace EventFlow.Commands
 {
     public interface ICommand
     {
+        Task<ISourceId> PublishAsync(ICommandBus commandBus, CancellationToken cancellationToken);
+        ISourceId GetSourceId();
     }
 
-    public interface ICommand<in TAggregate, out TIdentity> : ICommand
+    public interface ICommand<in TAggregate, out TIdentity, out TSourceIdentity> : ICommand
+        where TAggregate : IAggregateRoot<TIdentity>
+        where TIdentity : IIdentity
+        where TSourceIdentity : ISourceId
+    {
+        TIdentity AggregateId { get; }
+        TSourceIdentity SourceId { get; }
+    }
+
+    public interface ICommand<in TAggregate, out TIdentity> : ICommand<TAggregate, TIdentity, ISourceId>
         where TAggregate : IAggregateRoot<TIdentity>
         where TIdentity : IIdentity
     {
-        ISourceId SourceId { get; }
-        TIdentity AggregateId { get; }
     }
 }